requirements.txt
Django==1.9.6
django-filters==0.2.1
djangorestframework==3.3.3
flake8==2.5.4
mccabe==0.4.0
pep8==1.7.0
psycopg2==2.6.1
pyflakes==1.0.0
